extract exports from municipality pages to components folder
