+++ src/components/gemeente/intake-hospital-barscale.tsx
@@ -1,0 +1,33 @@
+import BarScale from '@/components/barScale';
+import { HospitalAdmissions } from '@/types/data.d';
+
+import siteText from '@/locale/index';
+const text: typeof siteText.gemeente_ziekenhuisopnames_per_dag =
+  siteText.gemeente_ziekenhuisopnames_per_dag;
+
+export function IntakeHospitalBarScale(props: {
+  data: HospitalAdmissions | undefined;
+  showAxis: boolean;
+}) {
+  const { data, showAxis } = props;
+
+  if (!data) return null;
+
+  return (
+    <BarScale
+      min={0}
+      max={100}
+      screenReaderText={text.screen_reader_graph_content}
+      value={data.last_value.moving_average_hospital}
+      id="opnames"
+      rangeKey="moving_average_hospital"
+      gradient={[
+        {
+          color: '#3391CC',
+          value: 0,
+        },
+      ]}
+      showAxis={showAxis}
+    />
+  );
+}
