+++ src/pages/gemeente/[code]/positief-geteste-mensen.tsx
@@ -1,4 +1,3 @@
-import BarScale from '@/components/barScale';
 import { FCWithLayout } from '@/components/layout';
 import { getMunicipalityLayout } from '@/components/layout/MunicipalityLayout';
 
@@ -6,6 +5,8 @@ import siteText from '@/locale/index';
 
 import { LineChart } from '@/components/charts/index';
 import { ContentHeader } from '@/components/layout/Content';
+
+import { PositivelyTestedPeopleBarScale } from '@/components/gemeente/positively-tested-people-barscale';
 
 import Getest from '@/assets/test.svg';
 import formatDecimal from '@/utils/formatNumber';
@@ -23,33 +24,6 @@ import MunicipalityLegenda from '@/components/chloropleth/legenda/MunicipalityLe
 
 const text: typeof siteText.gemeente_positief_geteste_personen =
   siteText.gemeente_positief_geteste_personen;
-
-export function PostivelyTestedPeopleBarScale(props: {
-  data: PositiveTestedPeople | undefined;
-  showAxis: boolean;
-}) {
-  const { data, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={10}
-      screenReaderText={text.screen_reader_graph_content}
-      value={data.last_value.infected_daily_increase}
-      id="positief"
-      rangeKey="infected_daily_increase"
-      gradient={[
-        {
-          color: '#3391CC',
-          value: 0,
-        },
-      ]}
-      showAxis={showAxis}
-    />
-  );
-}
 
 const PostivelyTestedPeople: FCWithLayout<IMunicipalityData> = (props) => {
   const { data } = props;
@@ -78,7 +52,7 @@ const PostivelyTestedPeople: FCWithLayout<IMunicipalityData> = (props) => {
           <h3>{text.barscale_titel}</h3>
 
           {positivelyTestedPeople && (
-            <PostivelyTestedPeopleBarScale
+            <PositivelyTestedPeopleBarScale
               data={positivelyTestedPeople}
               showAxis={true}
             />
